Clean up React Doctor findings and finish header refactor
- Remove unused devDependencies: baseline-browser-mapping, globals,
  typescript-eslint (still resolved transitively, so nothing breaks)
- Extract buttonVariants into button-variants.ts so button.tsx exports
  only components (Fast Refresh / only-export-components)
- Extract navigationItems into navigation-items.ts for the same reason,
  and point mobile-menu at the new module
- Wire mobile menu to useScrollToSection and use an Effect Event for the
  Escape handler so the keydown subscription doesn't re-run on every render
- Trim stack icons
