+++ package.json
@@ -44,17 +44,14 @@
     "@types/react-dom": "^19",
     "@typescript-eslint/eslint-plugin": "^8.47.0",
     "@typescript-eslint/parser": "^8.59.4",
-    "baseline-browser-mapping": "^2.9.19",
     "eslint": "^9.39.1",
     "eslint-config-next": "^16.2.6",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-react": "^7.37.5",
-    "globals": "^16.5.0",
     "postcss": "^8.5.14",
     "prettier": "^3.6.2",
     "prettier-plugin-tailwindcss": "^0.7.1",
     "tailwindcss": "^4.1.18",
-    "typescript": "^5",
-    "typescript-eslint": "^8.59.4"
+    "typescript": "^5"
   }
 }
