+++ src/components/header/navigation-items.ts
@@ -1,0 +1,5 @@
+export const navigationItems = [
+  { href: "#about", id: 3, label: "About" },
+  { href: "#stack", id: 2, label: "Tech Stack" },
+  { href: "#features", id: 1, label: "Features" },
+];
