+++ src/components/header/navigation.tsx
@@ -1,10 +1,6 @@
 import { scrollToSection } from "@/lib/utils";
 
-export const navigationItems = [
-  { href: "#about", id: 3, label: "About" },
-  { href: "#stack", id: 2, label: "Tech Stack" },
-  { href: "#features", id: 1, label: "Features" },
-];
+import { navigationItems } from "./navigation-items";
 
 /**
  * Main navigation component for header
